- Solves insertion sort erroneously not being called - Solves problems with the algorithm - Revises the row struct to accommodate the ambiguity between a double and a float, saving the precision correctly - Properly calls the division of the algorithms

+++ Ex1/headers/ReaderUtility.h
@@ -12,7 +12,7 @@ struct Row {
     int ID;
     char *Name;
     int Number1;
-    float Number2;
+    double Number2;
 };
 
 void free_row(struct Row *row);
